honeymoon/requirements.txt
pillow>=10.0.0
pillow-heif>=0.13.0
numpy>=1.24.0
scikit-learn>=1.3.0 